chore(types): delete old file and create new for field
